Clarify the API compatibility policy and improve the deprecations table.
In discussion around #2697, we realized the existing API compatiblity
policy is a little unclear, leading to confusion.
We discussed and clarified the intentions of this policy in #2790.
This change adds an example to the policy, and tweaks the deprecations
table to make a bit more sense.

Fixes https://github.com/tektoncd/pipeline/issues/2790