Add docs on creating a patch release
When making 0.13.1 I missed some items in the milestone and now need to
make 0.13.2 :( To help myself in the future I've added some docs to
explain what one should look for when making a patch release.

(Also I find docs like this really help me cuz I tend to jump in and out
of projects and by the time I make a release again I may have forgotten
what I need to do!)

I also removed "Update the new release in GitHub with the same links..."
from the official instructions because it looks like
https://github.com/tektoncd/plumbing/tree/master/tekton/resources/release#create-draft-release
does that for you.