Add bumping catalog dependency to process of creating a release
Eventually we want to be able to have Tasks declare what versions they
are compatible with and test against those
(https://github.com/tektoncd/catalog/issues/373) but in the meantime, it
seems reasonable to bump the version used in catalog tests every time we
do a release (and even once we have https://github.com/tektoncd/catalog/issues/373,
we would want to be running tests against new versions as well - and this
could have the benefit of giving us information about issues in a release
early, before users notice them!)

Question: how does the person making the release ensure that this bump
didn't make anything start failing? I think we should still make this
part of the process before we figure this out entirely, but we might
want to create a cron or something that runs all the tests nightly, and
maybe have the build cop trigger the release manually to ensure there
aren't any problems?