Don't use pointer when reading termination message
Using a pointer here caused us to get the warning "Implicit memory
aliasing in for loop" from gosec. Not to mention that functions with
side effects are hard to maintain and reason about over time, so
instead, we'll now have one function that gets the values to use and
we'll actually do the mutation outside the function, so no reason to
pass in a pointer.