Fix pipelinerun_taskrun_duration_seconds metric
This CL changes reporting taskrun duration as
part of pipelinerun metric to use knative
metrics Record API to match with the rest of the
existing metrics.

We have been using opensensus stats Record API for
reporting pipelinerun_taskrun_duration_seconds metric
as oppose to other metrics that we used knative metrics
API to report them. At this point only the metrics
that are reported with knative metrics are being
correctly plumbed to metrics backend. So as a result
pipelinerun_taskrun_duration_seconds was missing on
the metrics backend.