Handle dag in pipelineresoultion
In the pipelinerun controller, today we follow this logic:

- build the dag from the spec, using the dag module
- resolve the pipelinerun state using the spec and status, using the
  resources module
- get a list of candidate next tasks from the dag module, passing part
  of the state
- get a list of next tasks from the resources module, using the list
  of candidates and the pipeline run status

The separation of concerns between the dag, resources and reconciler
modules feels a bit mixed up.

This resolves part of the issue, by moving the invocation of the dag
building as well as obtaining the list of candidates to the dag
module, and aggregating the dag and the pipeline state into a new
struct ResolvedPipelineRun.