Define v1alpha1 Run type and generated scaffolding
This defines the basic Run type, which will be used to specify
executions of Custom Tasks and report status of those executions.

This change doesn't wire the new type in to any controllers or
validating webhooks, so that this time nothing will happen when a Run is
created. It also doesn't wire in to PipelineRun execution to enable
Pipelines to specify Custom Tasks.