Validate TaskRun compatibility with the Affinity Assistant
A TaskRun that mount more than one PVC-backed workspace is incompatible
with the Affinity Assistant. But there is no validation if the TaskRun
is compatible - so the TaskRun Pod is stuck with little information on why
to the user.

This commit adds validation of TaskRuns. When a TaskRun is associated with
an Affinity Assistant, it is checked that not more than one PVC workspace
is used - if so, the TaskRun will fail with a TaskRunValidationFailed condition.

Proposed in https://github.com/tektoncd/pipeline/issues/2829#issuecomment-645868118
Closes https://github.com/tektoncd/pipeline/issues/2864