update tutorial doc
This doc takes the current tutorial doc and updates it. The current git-repo google skaffold no longer holds the proper resources in the repository for the tutorial. As a user I have to use a different git repo for the tutorial. As a user I will use docker hub because using docker hub is easier than creating a local image registry