Update releases tasks/pipeline to use workspaces
Let's try to use workspaces to do the release instead of relying on
PipelineResource and v1alpha1 golang catalog resources.

This updates the tasks and pipeline to use workspaces.
This also update a little bit the readme with `kubectl diff` commands
to show what will be applied before applying it.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>