Makefile: add a cross target 🌵
This will allow building for different target (x86_64, arm, arm64,
s390x and ppc64le).

This doesn't mean we support those architecture (especially as `ko`
doesn't really support multi-arch for now). But it is nice to be able
to see if our code compiles and can be package (manually,
*unofficially*, for those target). A check will be added in plumbing
so that we validate that, at least, a PR doesn't break those.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>