Use `ko://` in e2e tests to exercices current code 🙃
Right now, some examples are exercicing old version of the code, and
definitely never the changes that would be on the images in a PR.
Using `ko://` instead helps with that (e2e tests are using `ko` to
apply the samples).

This link `.ko.yaml` in test folder to use the proper base image (as
`ko create …` seems to be running from the `test` folder).

Signed-off-by: Vincent Demeester <vincent@sbr.pm>