Dont write creds-init files if none of that type are mounted
Currently our creds-init code builds files on disk even when a credential
of a given type doesn't exist. For example, a .ssh/known_hosts file
is written even if there are no git ssh keys mountes in the step container.
It will be an empty file. The same is true for .docker/config.json and
.gitconfig.

This commit changes our creds init code to only write files for credentials
that have actually been mounted.