Document that Task Results dont get trimmed
The contents of task results are not processed by Tekton in any way - leading
and trailing whitespace characters are emitted verbatim from files written to
/tekton/results.

This commit adds documentation to make clear that Tekton does not process result
contents in any way.

In addition, I've added a short intro to the beginning of the results section
to give some focus on what results are useful for.