Flesh out the docs on using results in Pipelines
The existing documentation related to Results in Pipelines was
accurate but didn't go a long way to justify the feature's
existence. There was also a broken link to Task Results and
a slightly muddy description of how results affect Task ordering.

Here I've tried to clarify a little bit both the intended usage of
results in Pipelines and the technical details surrounding them.