tekton/release-pipeline: run test after precheck
We do a precheck to make sure we can do the release, and we usually
run the go build (and the actually `publish` task) after. But for some
reason, we do run tests as soon as the pipeline start. This changes
will start the go test task after the prechecks are done (so if
prechecks fails, we don't run them for nothing).

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>