Fix minor format issues
Some code blocks were not indented inside of ordered lists, resulting
in the lists' numbering being reset, so I indented them to match the
others in the list.

A 'json' code block was not fully valid json, so it was rendering with
a red `...` which was different than the rest of the page. The language
identifier was removed to match the other codeblocks in the doc.

The `finally` section was not referenced in the "definition" doc, so I
added it to make it more discoverable.