Use a hash of the workspace name in PVCs from template
Today, PVC name from volumeClaimTemplate is retrieved from
an optionally user submitted prefix in addition to the workspace name
and the owner name (PipelineRun name or TaskRun name).

The problem with this is that the workspace name need to follow the
naming requirements for Kubernetes resource names. It might also
be sensitive to very long names.

This commit only uses a fixed length hash of the workspace name
and the owner name (PipelineRun name or TaskRun name)

Closes #2941