Add URL to the ResourcesResults returned by resources.
Today, this contains commits and digests (for Git and Image Resources). These are great,
but need the addition of a URL to really verify or fetch the content again. This could be
fetched off the Resource itself, but that would require resolving the Resource (since it could
be passed by reference and change after the fact).

We decided not to dereference Resources into the Status because they are still alpha. This
gives us the requried information, without coupling the beta API to an alpha field.