Change podconvert.MakePod func into a configuration struct with methods
While working on a feature to optionally disable Tekton's built-in credentials
initialization process, it became apparent that the podconvert.MakePod func
would need an extra boolean argument in its signature. The signature is already
quite long which makes reading it difficult.

This commit separates the signature into configuration options and the task
spec / run. In a follow-up commit I plan to add the boolean flag described
above as another configuration option in the PodBuilder struct.