Remove mention of ssh-keygen from auth doc
The auth docs mention that tekton generates a known_hosts file using
ssh-keygen but this isn't quite right - we used to fetch the key
using ssh-keyscan but now accept it by updating git's core.sshCommand.

This PR updates the auth doc to describe the implementation we use to
populate known_hosts when its missing.