Fix TestTaskRunStatus test to run on s390x
TestTaskRunStatus test fails for s390x architecture because busybox
image in the code is specified with concrete SHA. SHA is unique for
each image and as a result for s390x case the amd64 image with the SHA
is pulled. Test failed.

Adding s390x specific SHA to s390x build solves this problem.

Signed-off-by: Yulia Gaponenko <yulia.gaponenko1@de.ibm.com>