Fix small typo in install doc for default workspace config
When defining a default taskrun workspace in the config-defaults YAML
file, the user should start their workspace configuration with the
`|` character, to indicate that it's a block string. This character
is currently missing from the example in the install doc.

This commit adds the trailing `|` character to the config-defaults
example in the install doc.