Added `-stdout_path` and `-stderr_path` flags to entrypoint.
When `-stdout_path` and `-stderr_path` are set, the entrypoint binary will copy
stdout and stderr to the specified paths. This can be used to capture the stdout
or stderr of a step in task results.

The default values of these flags are obtained from the `TEKTON_STDOUT_PATH` and
`TEKTON_STDERR_PATH` environment variables, so users can specify these variables
in the step's environment to enable stdout/stderr capturing.