Adds support for metadata in TaskRef
Following the same behavior os TaskSpec which allow to specify Labels and
Annotations to be copied to TaskRun, this commit adds TaskRunMetadata to
TaskRef and the necessary code to copy the values.

This ensures a consistent behavior allowing to parameterize thes metadata and
allows them to be consumed by tooling.