Add nop-image to release instructions
In #3025 the nop-image was added to our release pipeline, this commit
updates our README to include an example that sets it.

Also removed the reference to resources.yaml from the instructions b/c
imo opening that file and modifying it is more confusing than helpful;
feels like it makes more sense to create a new PipelineResource instance
without bothering with that file.