Fix version annotation on created pods 👇
The `ldflags` doesn't work with `ko`, so we end up annotating all our
pods with `devel` as a version (`pipeline.tekton.dev/release: devel`).

This aims to fix it by adding a flag in the controller (`-version`)
and passing it in the yaml. The release task script can then replace
`devel` by the correct version (as what's already done).

This also remove the `ldgflags` part.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>