Rename "Artifact Storage" to "PipelineResource Storage" in install docs
The install docs aren't super clear about the meaning and purpose of "artifact storage".
Reading the existing doc it would be understandable if a user mistakenly
thought that Tekton relies on the storage configuration for all data in
and out of Tasks. In fact this storage is specific to PipelineResources
that are linked "from" one to another in a Pipeline.

This commit updates the docs to drop mention of "artifact storage" and
instead directly reference PipelineResources, with supporting links and
a note again warning of the alpha-ness of the feature.