resolve params in when expressions
Instead of creating a new empty array and appending resolved when
expression, creating an array based on the existing one and replacing
in place the resolved when expression.

Also, do not evaluate when expressions until all parents are done,
wait for parent tasks to be completed before evaluating when expressions.
This check will make sure that tasks resulting in when expressions
failures, are not declared skipped in advance before their parents are done.

With when expressions having reference to task results, evaluate and apply
task results before checking if its to run finally tasks. This change is
necessary to make sure all the leaf nodes of DAG have resolved task results in
their when expressions to evaluate whether DAG is done or not. This change
of order works today since finally tasks does not need to resolve task results
yet. Once we add task results support in finally, we will have to make one
more call to apply task results after adding finally tasks to execution queue.