Sort pod container statuses based on Step order in taskSpec
This commit closes #3239

Tekton determines the TaskRun status message of a failed TaskRun based
on the results of the first terminated Step (pod container). Until now,
Tekton sorted pod container statuses based on the FinishedAt and
StartedAt timestamps set by Kubernetes. Occasionally, a Step terminated
in response to the first terminated Step could have the same timestamps
as the first terminated Step. Therefore, Tekton was not always able to
correctly determine what the first terminated Step was, and as a result,
Tekton may set an incorrect TaskRun status message.

In this commit, pod container statuses are sorted based on the container
order as specified by Tekton in the podSpec. Tekton bases this order on
the user provided taskSpec and Steps added internally by Tekton.
Therefore, Tekton accounts for internally added Steps when
sorting pod container statuses.