Update Release Cheet Sheat, remove instructions from README
The release-cheat-sheet.md file covers a lot of the same ground for official releases
as the README does, but the cheat sheet is more succinct. The release cheat sheet does
not currently include the additional nop image flag.

This commit adds the new nop built image flag and removes the official release section
from tekton/README in favor of linking directly to the cheat sheet.